Tesseract interface - don't overrun Unicode return buffer
Bug #708789 "10.06.0rc1: PDFwrite OCR produces corrupted files with certain settings"

We pass in the maximum size of the buffer and we should never write
past the end of that buffer, even if Tesseract returns more Unicode
code points than will fit.

The do..while loop could overrun the buffer by 1 (because we don't
check max_chars until the end of the loop). This is bad enough, but if
we continue on and get another iterator from Tesserat then that could
overrun by 'count' bytes before we detect it at the end of the loop.

Make sure count does not exceed the size of the buffer before we do
the UTF8 to Unicode conversion.

While passing I noted that there's a typo in the parameter code for the
'Always' OCR setting, fix that too.