graphics library - prevent page sizes of 0
Bug #708800 "Heap-buffer-overwrite when using lips4 output device"

The problem is that the PDF file sets a media width of 0.12 which, at
200 dpi, works out to less than 1 pixel.

There's nothing wrong with the patch supplied, but it is specific to
the lips4 device so it dosn't prevent the same, or similar, problems
in other devices. And we have a lot of devices.

So in this commit, check the device width and height to make sure they
are at least 1 pixel, and check the resolution to make sure it is
greater than 0 in each dimension. We do this in gdev_prn_output_page_aux()
so that any device derived from gdevprn will then get the same
protection.