graphics library - don't delete files in %ram% on closefile
Bug #708810 "[BIGSLEEP-443268749] heap-use-after-free in ramfile_close()"

The basic problem is that close file shouldn't delete files and
deletefile shouldn't delete files which are open.

Define a new error to report attempts to delete open files (not really
needed, but in the spirit of the code).

Alter the 'refcount' member so that it is 1 on opening a file (not 2).

Don't count down the refcount when actually deleting a file.

Check the refcount in 'unlink' and return an error if it isn't 0.

Don't attempt to delete the file in _close().