XPS interpreter - bounds check POST table parsing
Bug #708812 "Heap-buffer-overread in xps_true_callback_glyph_name with a crafted XPS file"

The code does have a check for the 'glyph_name_index'th string
overrunning the buffer, but it doesn't check intermeditate strings.

In this case the cause is that glyph_name_index is unfeasibly large, but
corruption of the pascal string data could lead to the same effect so
its better to check the pointer in the loop which fixes both potential
problems.

Also, if 'glyph' is sufficiently large the pointer to retrieve the
glyph_index could lie past the end of the POST table (technically this
shouldn't be possilbe, but we're dealing with damaged data here). So
check that too.