graphics library - cater fo subclassing on device open method
Working through new code for a customer, and this came up. The
setpagdevice functionality ends up calling the PostScript 'opendevice'
operator, and the C implementation didn't know about subclassing
devices, so it only set the 'is_open' member on the top device.

The put_params code then later copied the member from the bottom device
to the top (because devices can be closed unexpectedly) which caused the
device to be regarded as closed.

Fix is simply to set the 'is_open' member to true in the entire chain,
it should be the same for all device in the chain.