pdfwrite - close the correct device on OwnerPassword change
Bug #708835 "heap-use-after-free read in pdf_page_id()"

The correct device to close is the original, not the copy. The reason
for closing the device is to write streams with the correct password
for the security method. This is why we check that the passwortd has
changed, and no pages have yet been written, before allowing the new
password.