Fix psd devices use of downscaler.
We were counting the number of colors that we needed to output,
and initialising the downscaler to that number of planes. This
works fine for PDF files, but can go wrong for postscript ones,
especially where tags are involved.

For postscript files, we might only have CMYK + Tags, but the
color setup will be that we have 16 colors, with tags in position
15.

When we call the downscaler to 'getbits' it would only fetch the
first 5 colors, and hence the tags value would be garbage.

Instead fetch all the planes it can give us, and we'll pick the
correct ones to output.