Bug 708841: Avoid overflow resizing a ramfs file
We already error for an attempt to resize a ram file to a size that would
overflow a 32 bit integer.

In this case, the request is for a value close to the limit, which ends up
overflowing the signed integer value as we round up to calculate the number of
ramfs blocks the request requires.

Recast that calculation into an unsigned int, after dividing down to get to a
number of ramfs blocks, recast to a signed int. That avoids the overflow in the
interim value.