Bug 708847: Avoid dereferencing NULL pointers C0/C1 in Type 2 functions
The problem is there is a disconnect bertween an optimisation that relies on
the default values for the C0/C1 arrays, and the requirement to scale those
values.

Even if the original function relies on the default values (meaning we can avoid
creating the arrays), if we are asked to scale those values, we *must*
have arrays to store those scaled values.

Note: I've been unable to find a case where this code gets called, which probably
explains why this problem has never appeared in real use.