Bug 708850 preparation: Copy tags setup into pdf14 device.
This was found as part of the work on bug 708850, but it's probably
the right thing to do regardless of whether that bug is fixed or
not.

When we are looking for a prototype upon which to base the
pdf14 device, previously we didn't copy the tag information early
enough so that it was reflected in the color_info for the
device prototype.

This could lead to us incorrectly adding extra planes to the
device in the belief that we hadn't allowed for tags already.