Bug 708850: Fix putimage to a device with an nchannel profile.
On 'pushing' the transparency compositor, we have spotted that
the underlying device is using an nchannel profile, and replaced
that with a cmyk one.

Accordingly, when we putimage, the code correctly takes the
CMYK values in the buffers and converts them back to nchannel
values.

Unfortunately, if the nchannel stuff has n > 4, then it leaves
the n-4 extra channels in the data anyway, meaning that we
end up with all channels >= n being shifted up by n-4.

The fix, implemented here, is to throw away the extra channels
from the pdf14 buffers (which should be entirely empty anyway).