pdfwrite - Use correct limit when sampling DeviceN alternate space
Bug #708834 "Possible out-of-bounds array access in `convert_DeviceN_alternate()"

The code in question explores the space of a DeviceN alternate function
by setting each of the components to 0 and 1 in every possible
combination, then converting that to the alternate space. We later use
the sampled data as the input to a sampled data function to create an
'equivalent' to the original tint transform but targeting a different
base space.

The error is simply that the 'cascade' which controls the generation of
the input values needs to be limited (obviouosly) to the number of
components in the input space; the code was using the number of
actual samples, which is 2^components.