PS interpreter - give a more helpful error on TBCP selection from PJL
Bug #708844 "ps2pdf fails to decode Lexmark PS3 postscript file with /undefined in --filter--"

We removed the TBCPDecode filter because it isn't a standard part of the
PostScript specification. However, the PJL language, when switching to
PostScript when the pritner supports both PCL and PostScript, can state
that the PostScript is encoded using the TBCP communications protocol.

Change the code to generate an error immediately and give a more
helpful message.