Bug 708854: Avoid uninit pointers with broken PDF collections
In the event of an error retrieving the names and contents of files in a PDF
collection, string pointers for the file name and/or scratch file could end
up uninitialised, and then the calling code attempt to dereference them.

Make sure the pointers are properly initialised, and that the calling code
copes with them being null.