Ensure we free all the strings stored for a PDF collection
For each embedded file in a PDF collection we store two strings, so the array
of string pointers is twice as long as the number of embedded files.

The loop counter limit needs to be twice the number of embedded files to free
all the strings.

Noticed in passing when fixing Bug 708854