graphics library - avoid NULL dereference from clist
Bug #708851 "Crash on unknown address read when using the bmpsep1 output device"

The halftone load method, when called from the clist, has a NULL
pointer to the graphcis state (there is no real graphics state in the
clist I believe). This causes a seg fault with some code which tries to
check that the halftone is 'valid' for a type of object.

Basically this is error detection code. I don't believe it is possible
to trigger this on the back of a clist; we can't still be trying to use
a restored-away halftone in this condition.

So we can avoid the problem by simply not performing the check if the
pointer to the graphics state is NULL.

There's a second dereference, which again I don't believe it should be
possible to hit from the clist, but for safety check that pgs is not
NULL before dereferencing it. If it is exit with an error.