graphics library - avoid low address dereference from clist
Bug #708851 "Crash on unknown address read when using the bmpsep1 output device"

This is the second of two bug fixes required for this report. This one
can't happen unless the earlier commit has already been implemented.

The 'plane extract' device, apparently used to image a single plane for
separated output, doesn't implement a fill_stroke_path method. When
used with the bmpsep device, and a clist, this causes the device to
fall through to the default handler, which simply punts it on to the
next device.

In this case, we eventually reach the bmpsep1 device, which has been
reconfigured as a clist device. Because we are actually executing a
clist at this point, the device is a clist reader, but the method is
expecting a clist *writer*.

I'm not completely certain that this is correct, however inspection of
the plane extract device indicates that, in general, it implements
methods for all the rendering cases. So the lack of a fill_stroke_path
method is almost certainly an oversight.

In this commit simply add a method which does nothing more than call
the existing fill method, and then the stroke method.