Bug 708850 Continued: Fix calculation of planes to drop in putimage.
The previous commit for this bug introduced the idea that we could
drop some planes during the putimage if our output profile is an
NChannel one.

It turns out that the calculation for 'how many planes to drop' was
incorrect in the case where we have a top level blend space that's
not the same as the device.

This commit should address that.