Attempt to fix indeterminisms/valgrind warnings with devn colours.
A gx_device_color that represents a device n color holds up to
GX_DEVICE_COLOR_MAX_COMPONENTS component values.

It holds no indication of how many of these are actually used
though, so when we compare 2 such values, we have to compare all
the components. See gx_dc_devn_equal.

Unfortunately, in several places we create devn colors without
setting all the components to a known value; this results in
valgrind warnings and possibly indeterminisms in behaviour.

The fix, implemented here, is to always set all the unused
components to zero when creating such colors.