Bug 708876: Fix compositor device lifespan
In some cases, from gx_forward_composite() the compositor device is returned to
the caller for further handling. In other (most) cases, the compositor becomes
the target for the forwarding device, and the reference should not be
propagated by the caller.

Firstly, that ends up with a reference count one too high, since the local
reference doesn't persist, so fix that.

Also, enforce the non-propagation by setting the returning pointer to NULL.