Bug 708882/708883: Fix saving/restoring color_info in transparency.
First issue: The transparency compositor hadn't been updated to reflect a change
in the device API. Rather than device procs being pickled into the device
prototype declaration, we now use a single function pointer
(initialize_device_procs) to populate the device procs. Meaning that simply
copying the procs from the prototype just results is all NULLs now.

So, call through initialize_device_procs to set the procs before copying them.

Secondly, when dealing with a clist morphed output device, we impose the
color_info required for the transparency group onto the clist device, and save
the original color_info to be restored later.

Problem was, we did that twice: once in gs_pdf14_clist_device_push() then again
in pdf14_clist_composite(), meaning we overwrote the saved color_info with
the modified, and ultimately restored the wrong color_info back to the clist
device.

Ultimately meaning, the output device could potentially render to the the wrong
colour setup - in this case, rendering to 24bpp when the device expects
8bpp.