clipping - fix a bug with pdfwrite
This is from a customer file, so no bug report. Commit
570688bfdfba4b1b2d9fee3b454ff733fe3ffa89 fixed an issue with pdfwrite
where clipping was not being handled correctly when the corners of a
new rectangular clip were the same as the existing clip.

However, this missed a small but important point; we later use the
'changed' variable to decide whether to use the new path as is, or
regenerate it. By altering the value of 'changed' I changed the
conditions for that test.

Essentially there are three results here
1) The new clip is larger than (r equal to) the old clip, ignore the new
one.
2) The new clip is entirely contained inside the old clip, use the new
one 'as is'.
3) The new clip and old clip partially overlap. The result is a new
rectangular area (and path). We defer creating the path.

We were messing up case 3 because manipaulting 'changed' for the benefit
of detecting co-incident points caused us to take the wrong action.

This commit separates out the check for con-incidence, and changes
caused by intersecting the areas. We modify 'same' to detect if all
four corners of the two clip rectangles are either the same, even if
we changed some of the ones in the 'new' clip to be the same as the
existing clip.

If all four corners are the same, after any modifications, as the
original clip then we can ignore the new clip.

If none of the corners of the new clip have changed, then we can just
use the new clip directly.

Otherwise we have created an intersection, and defer creating the path.