Contrib devices - fix image fallback handling
Bug 708865 "Null pointer dereference when using the lips4v output device"

The problem here is, simply, that the 'fallback' handling in the lips4v
device simply doesn't work. At all. It looks likely that it never has,
it certainly hasn't worked since the introduction of type 3 and 4 images
in PostScript language level 3.

Hardly surprising since the copyright date of the device is 1999, the
same year level 3 was published.

There is code that purports to handle type 3 and 4 images, by falling
back and letting the graphics library handle them, then catching the
data at a lower level (copy_mono and copy_color), but the type 3 image
code requires a 'planes_wanted' in the image enumerator (which the device
doesn't provide) and the code didn't copy the relevant 'common' portion
of the fallback enumerator to its own custom enumerator, and that is
used by the interpreter, which falls in a heap because the data isn't
initialised.

So this is actually fairly simple. After creating the fallback
enumerator (which we store in the vector device's 'default_info') we copy
the common portion from that into the custom enumerator. That means the
interpreter now has sensible data to work with.

We create 'planes_wanted' and 'flush' methods for the custom enumerator
'procs'. These simply pass on the call to the fallback enumerator, the
'image_plane_data' and 'end_image' methods already do this. This means
that the graphics library now has the required methods available.

This did require a small change in the vector device implementation, so
that we don't try and use a non-existent 'bbox_info'.

Desk-checking it looks very much like the Open Printing device has
exactly the same problem, as does the 'escv' device(s). So updated
those similarly.

This cures the crash in the lips4v device, but I have no way to
determine if the output is reasonable. The same for the escv and
opvp devices.