Device interface - make future changes easier.
The PDF14 transparency compositor needs to create new rendering devices
during the course of its operation. We *don't* want those devices to
be created along with the entire pipeline of subclassing devices, so
we set the various 'Pushed' flags for the internal devices to 'true'
so that opening the device doesn't try to push them again.

But its awkward having to do this separately, and easy to forget, which
causes crashes.

So make a central routine in the same file as all the other 'kernel'
internal dvice code to do the work, and call that from the PDF14
compositor instead of doing the work there.