Bug 708930: Ensure buffer space for call into libjpeg
libjpeg doesn't allow writing to be suspended when writing out a marker, so
there has to be enough write buffer available to complete the operation.

This isn't a regression, exactly, it's just as things change, the problem can
be unpredictably exposed or obscured.

Before calling into the libjpeg code, we have to make sure that the write buffer
for the target stream has at least 700 bytes available.