Bug 708608: Fix confusion caused by modern C having a boolean type
Because C99 and later have a proper, built-in boolean type, rather than a
typedef of an underlying integer type, the check for whether we're rendering
an image into a pattern accumulator was yielding an unexpected result.

The default dev_spec_op method returns an "undefined" error when queried with the
gxdso_in_pattern_accumulator op, that is, a negative integer. When "bool" was
simply a typdefed integer value, that was fine, and the following lines
checked for and handled that.

With an actual boolean type, false == 0, and true == !false, so any value other
than 0 became true.

That then changed the decision on whether to gridfit images, meaning the user
space to device space rounding was handled differently. That then caused a
scanline of the image to be repeated.

The value is now assigned based on a conditional, so it correctly assigns a
true/false value.