pdfwrite - add clip to 'viewer' graphics state
Bug #708954 "Clipping incorrect in output PDF file"

The pdfwrite device wasn't tracking the clip as part of the graphics
state, just as an element in the device. The file here sets a clip as
part of writing out a form, it then grestores back and writes the same
(or very nearly the same) clip before emitting text.

The text should be clipped, but the pdfwrite device doesn't emit the
second clip, because it thinks the prior one is still in force. Because
we weren't tracking it in the graphics state.

This commit adds the clip path and id to the 'viewer' graphics state, so
that it changes as we save and restore the output graphics state.
We need to free the paths on exit if the graphics state is not empty and
we need to set them to be empty when creating a viewer state from the
graphics state (for things like capturing type 3 font CharProcs).

The pdf_update_alpha() function was saving the viewer state
inappropriately, and not restoring it back again, which led to the
problem persisting because the output graphics state didn't match the
'viewer' graphics state.