Bug 708956: Fix flushing the final buffer of JPEG passthrough data
The jpeg_finish_decompress() call to libjpeg is what prompts the library to
(potentially) consume its final bytes of input. That then updates the
"next_input_byte" pointer to end of the data (after the EOI), that's what we
base the last real call to the passthrough function pointer on.

So we need to delay passing that final buffer to the passthrough function
pointer until after we've returned from gs_jpeg_finish_decompress() to be
certain that the EOI marker will be included in the passed through data.