PDF interpreter - treat broken Info dictionary as a missing dictionary
This is from a customer report with a private PDF file so no bug
number.

The file has been created with the Info and Root objects merged into a
single dictionary. This causes an error in the PDF interpreter when we
try to enumerate the Info dictionary keys, because the dictionary
contains a reference to the Pages tree (which an Info dictionary should
not contain) and that tree has circular references.

Fortunately we can treat a broken Info dictionary as being a missing
Info dictionary (since it is optional) and carry on.