Bug 708897: Fix assert condition
There is one condition where the pointer for the current position of the PS
exec stack can sensibly be below the bottom address of the stack when we call an
operator, and that is in the extremely rare circumstance where we are calling
the interal pseudo-operator that triggers an ExecStackUnderflow error.

This is essentially a last line of defence situation, since the underflow
condition is mostly caught elsewhere.

The assert condition needs to cope with the specific situation.